Add shell script for converting markdown to latex
This commit adds a shell script that converts a markdown file to a LaTeX file using the pandoc tool. The script can be executed by running the command "pandoc markdown_file.md -o latex_file.tex". This addition enhances the functionality of the project by providing an automated way to convert markdown files to LaTeX format.